update manage event
